Add script to check a line-by-line's completeness
The script in question iterates along its arguments to make sure that
(most of) the lines in the one are in the other. That way, changes to
files that I want to do line-by-lines of are pointed out to me. Still a
WIP, but I hope to automate this process, so that _all_ of the files I
do a line-by-line of get the treatment whenever I build the site (or
something similar).