Add auto-generated CV page
The page iterates along the CV categories in _data/cvcategories.yml, and
looks for pages throughout the site that have cvinfo in their front
matter, then inserts those details to the page in a pleasant format.

Voila, with that I can add a page depicting a project to the site
(anywhere!) and it will get automagically added to the CV page.